+++ packages/apify/src/typedefs.ts
@@ -34,10 +34,12 @@ export interface ActorRunWithOutput extends ActorRun {
     output?: Dictionary | null;
 }
 
+/** @ignore */
 export function entries<T extends {}>(obj: T) {
     return Object.entries(obj) as [keyof T, T[keyof T]][];
 }
 
+/** @ignore */
 export function keys<T extends {}>(obj: T) {
     return Object.keys(obj) as (keyof T)[];
 }